Increase memory constants for state layers and trace reexecution
The memory constants for both 'TriesInMemory' in state database and 'defaultTraceReexec' in tracers API have been increased from 128 to 512. This adjustment was made to enhance performance and allows more layers to be kept in RAM, and increases the number of blocks the tracer can reexecute to produce missing historical state.